+++ test/samples/destructuring.js
@@ -29,7 +29,7 @@ module.exports = [
 		output: `
 			function pythag ( ref ) {
 				var x = ref.x;
-				var ref_y = ref.y, z = ref_y === void 0 ? 1 : ref_y;
+				var z = ref.y; if ( z === void 0 ) z = 1;
 
 				return Math.sqrt( x * x + z * z );
 			}`
@@ -80,22 +80,10 @@ module.exports = [
 		output: `
 			function pythag ( ref ) {
 				var x = ref[0];
-				var ref_1 = ref[1], z = ref_1 === void 0 ? 1 : ref_1;
+				var z = ref[1]; if ( z === void 0 ) z = 1;
 
 				return Math.sqrt( x * x + z * z );
 			}`
-	},
-
-	{
-		description: 'disallows compound destructuring in declarations',
-		input: `var { a: { b: c } } = d;`,
-		error: /Compound destructuring is not supported/
-	},
-
-	{
-		description: 'disallows compound destructuring in parameters',
-		input: `function foo ( { a: { b: c } } ) {}`,
-		error: /Compound destructuring is not supported/
 	},
 
 	{
@@ -217,7 +205,7 @@ module.exports = [
 
 		output: `
 			var value = obj.name;
-			var obj_description = obj.description, description = obj_description === void 0 ? null : obj_description;
+			var description = obj.description; if ( description === void 0 ) description = null;
 			console.log( value, description );`
 	},
 
@@ -232,8 +220,8 @@ module.exports = [
 		output: `
 			function foo (ref) {
 				if ( ref === void 0 ) ref = {};
-				var ref_arg1 = ref.arg1, arg1 = ref_arg1 === void 0 ? 123 : ref_arg1;
-				var ref_arg2 = ref.arg2, arg2 = ref_arg2 === void 0 ? 456 : ref_arg2;
+				var arg1 = ref.arg1; if ( arg1 === void 0 ) arg1 = 123;
+				var arg2 = ref.arg2; if ( arg2 === void 0 ) arg2 = 456;
 
 				console.log( arg1, arg2 );
 			}`
@@ -258,6 +246,84 @@ module.exports = [
 
 					console.log(element);
 				};
+			}`
+	},
+
+	{
+		description: 'deep matching with object patterns',
+
+		input: `
+			var { a: { b: c }, d: { e: f, g: h = 1 } } = x;`,
+
+		output: `
+			var c = x.a.b;
+			var x_d = x.d;
+			var f = x_d.e;
+			var h = x_d.g; if ( h === void 0 ) h = 1;`
+	},
+
+	{
+		description: 'deep matching with object patterns and reference',
+
+		input: `
+			var { a: { b: c }, d: { e: f, g: h } } = x();`,
+
+		output: `
+			var ref = x();
+			var c = ref.a.b;
+			var ref_d = ref.d;
+			var f = ref_d.e;
+			var h = ref_d.g;`
+	},
+
+	{
+		description: 'deep matching with array patterns',
+
+		input: `
+			var [[[a]], [[b, c = 1]]] = x;`,
+
+		output: `
+			var a = x[0][0][0];
+			var x_1_0 = x[1][0];
+			var b = x_1_0[0];
+			var c = x_1_0[1]; if ( c === void 0 ) c = 1;`
+	},
+
+	{
+		description: 'deep matching with sparse array',
+
+		input: `
+			function foo ( [[[,x = 3] = []] = []] = [] ) {
+				console.log( x );
+			}`,
+
+		output: `
+			function foo ( ref ) {
+				if ( ref === void 0 ) ref = [];
+				var ref_0 = ref[0]; if ( ref_0 === void 0 ) ref_0 = [];
+				var ref_0_0 = ref_0[0]; if ( ref_0_0 === void 0 ) ref_0_0 = [];
+				var x = ref_0_0[1]; if ( x === void 0 ) x = 3;
+
+				console.log( x );
+			}`
+	},
+
+	{
+		description: 'deep matching in parameters',
+
+		input: `
+			function foo ({ a: { b: c }, d: { e: f, g: h } }) {
+				console.log( c, f, h );
+			}`,
+
+		output: `
+			function foo (ref) {
+				var c = ref.a.b;
+				var ref_d = ref.d;
+				var f = ref_d.e;
+				var h = ref_d.g;
+
+				console.log( c, f, h );
 			}`
 	}